news_lab/backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
pydantic==2.11.7
pydantic-settings==2.10.1
requests==2.32.4
beautifulsoup4==4.13.4
feedparser==6.0.11
openai==1.102.0
httpx==0.28.1
redis==6.4.0
PyYAML==6.0.2
python-dotenv==1.1.1
pytest==8.4.1
pytest-asyncio==1.1.0
